+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/booking/contract/service/DefaultOccupancySeatRangeFinder.java
@@ -2,9 +2,9 @@ package com.seatliberator.seatliberator.reservation.application.booking.contract
 
 import com.seatliberator.seatliberator.reservation.application.booking.contract.OccupancySeatRangeFinder;
 import com.seatliberator.seatliberator.reservation.application.booking.model.ReservationOccupancyPolicy;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.ReservationReader;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.criteria.ReservationFilter;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.criteria.ReservationSeatOverlapCriteria;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.ReservationReader;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.criteria.ReservationFilter;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.criteria.ReservationSeatOverlapCriteria;
 import com.seatliberator.seatliberator.reservation.domain.reservation.Reservation;
 import com.seatliberator.seatliberator.reservation.domain.shared.SeatLocator;
 import com.seatliberator.seatliberator.reservation.domain.shared.temporal.InstantRange;
