+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/application/booking/contract/service/DefaultReservationPolicyCheckerTest.java
@@ -1,9 +1,9 @@
 package com.seatliberator.seatliberator.reservation.application.booking.contract.service;
 
 import com.seatliberator.seatliberator.reservation.application.booking.contract.result.ReservationRejectReason;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.ReservationReader;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.criteria.ReservationFilter;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.criteria.ReservationSeatLookupCriteria;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.ReservationReader;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.criteria.ReservationFilter;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.criteria.ReservationSeatLookupCriteria;
 import com.seatliberator.seatliberator.reservation.domain.reservation.ReservationStatus;
 import com.seatliberator.seatliberator.reservation.domain.shared.SeatLocatorFixture;
 import com.seatliberator.seatliberator.reservation.domain.shared.temporal.InstantRangeFixture;
