+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/application/booking/criteria/ReservationRangeOverlapCriteriaTest.java
@@ -1,7 +1,7 @@
 package com.seatliberator.seatliberator.reservation.application.booking.criteria;
 
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.criteria.ReservationFilter;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.criteria.ReservationRangeOverlapCriteria;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.criteria.ReservationFilter;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.criteria.ReservationRangeOverlapCriteria;
 import com.seatliberator.seatliberator.reservation.domain.shared.temporal.InstantRangeFixture;
 import com.seatliberator.seatliberator.reservation.domain.shared.temporal.SimpleInstantRange;
 import org.junit.jupiter.api.DisplayName;
