+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/application/booking/service/AbstractReservationCommandServiceTest.java
@@ -3,8 +3,8 @@ package com.seatliberator.seatliberator.reservation.application.booking.service;
 import com.seatliberator.seatliberator.reservation.application.booking.contract.ReservationCreateAuthorizedPolicy;
 import com.seatliberator.seatliberator.reservation.application.booking.contract.ReservationCreatePolicy;
 import com.seatliberator.seatliberator.reservation.application.booking.contract.ReservationCreator;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.ReservationReader;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.ReservationStore;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.ReservationReader;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.ReservationStore;
 import com.seatliberator.seatliberator.reservation.application.seat.port.out.SeatStore;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
