+++ reservation/reservation-persistence/src/test/java/com/seatliberator/seatliberator/reservation/persistence/book/jpa/JpaReservationPersistenceAdapterTest.java
@@ -1,8 +1,8 @@
 package com.seatliberator.seatliberator.reservation.persistence.book.jpa;
 
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.ReservationReader;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.ReservationStore;
-import com.seatliberator.seatliberator.reservation.application.booking.port.out.criteria.*;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.ReservationReader;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.ReservationStore;
+import com.seatliberator.seatliberator.reservation.application.reservation.port.out.criteria.*;
 import com.seatliberator.seatliberator.reservation.domain.reservation.Reservation;
 import com.seatliberator.seatliberator.reservation.persistence.AbstractPersistenceAdapterTest;
 import com.seatliberator.seatliberator.reservation.persistence.book.jpa.repository.ReservationRepository;
